requirements.txt
asgiref==3.7.2
astroid==2.15.5
async-timeout==4.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
dill==0.3.6
dj-rest-auth==4.0.0
Django==4.2.2
django-allauth==0.52.0
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
oauthlib==3.2.2
platformdirs==3.8.1
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
pylint==2.17.4
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.7.0
urllib3==2.0.3
wrapt==1.15.0
